feat: add setup for controller framework
