+++ controller4fiware/__init__.py
@@ -1,0 +1,1 @@
+from .Controller import Controller4Fiware
